requirements.txt
pandas
numpy
jupyter
SQLAlchemy
python-dotenv
psycopg2
openai
requests
pprintjson
tiktoken
tenacity
termcolor